+++ packages/apps/ability-hyperliquid/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 3.0.4 (2026-01-12)
+
+### 🧱 Updated Dependencies
+
+- Updated e2e-test-utils to 2.0.3
+
+## 3.0.3 (2026-01-10)
+
+### 🧱 Updated Dependencies
+
+- Updated ability-sdk to 2.4.2
+- Updated app-sdk to 2.5.3
+- Updated contracts-sdk to 7.1.0
+- Updated e2e-test-utils to 2.0.2
+
 ## 3.0.2 (2026-01-08)
 
 ### 🧱 Updated Dependencies